assistant_windows/requirements.txt
# Assistant Windows - Python Dependencies

# GUI Framework
PySide6>=6.6.0

# Global Hotkeys
keyboard>=0.13.5

# Screen Capture (fast - Windows DirectX)
dxcam>=0.0.5

# Screen Capture (fallback - cross-platform)
mss>=9.0.1

# Image Processing
Pillow>=10.0.0

# HTTP Client
requests>=2.31.0